requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
dnspython==2.3.0
frozenlist==1.3.3
gensim==4.3.1
google-search-results==2.4.2
idna==3.4
joblib==1.2.0
langchain==0.0.157
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
np==1.0.2
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
SQLAlchemy==2.0.12
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
